+++ common/src/main/java/pro/mikey/justhammers/recipe/HammerRecipes.java
@@ -2,6 +2,7 @@ package pro.mikey.justhammers.recipe;
 
 import dev.architectury.registry.registries.DeferredRegister;
 import dev.architectury.registry.registries.RegistrySupplier;
+import net.minecraft.core.Registry;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.resources.ResourceLocation;
@@ -14,7 +15,7 @@ public class HammerRecipes {
     public static final DeferredRegister<RecipeType<?>> RECIPE_TYPES = DeferredRegister.create(Hammers.MOD_ID, Registries.RECIPE_TYPE);
     public static final DeferredRegister<RecipeSerializer<?>> RECIPE_SERIALIZER = DeferredRegister.create(Hammers.MOD_ID, Registries.RECIPE_SERIALIZER);
 
-    public static final RegistrySupplier<RecipeType<RepairRecipe>> REPAIR_RECIPE = RECIPE_TYPES.register("repair", () -> RecipeType.register(ResourceLocation.fromNamespaceAndPath(Hammers.MOD_ID, "hammer_repair").toString()));
+    public static final RegistrySupplier<RecipeType<RepairRecipe>> REPAIR_RECIPE = RECIPE_TYPES.register("hammer_repair", () -> new RecipeType<>() {});
     public static final RegistrySupplier<RecipeSerializer<RepairRecipe>> REPAIR_RECIPE_SERIALIZER = RECIPE_SERIALIZER.register("repair", () -> new SimpleCraftingRecipeSerializer<>(RepairRecipe::new));
 
     public static void init() {