+++ neoforge/src/main/java/pro/mikey/justhammers/neoforge/HammersNeoForge.java
@@ -6,7 +6,6 @@ import pro.mikey.justhammers.Hammers;
 @Mod(Hammers.MOD_ID)
 public class HammersNeoForge {
     public HammersNeoForge() {
-//        EventBuses.registerModEventBus(Hammers.MOD_ID, FMLJavaModLoadingContext.get().getModEventBus());
         Hammers.init();
     }
 }